Refactor due to changes in incidence2 and trending
{i2extras} now explicitly 'Depends' on the upstream
[{incidence2}](https://CRAN.R-project.org/package=incidence2) package and
enforces a minimum version requirement upon this (i.e. incidence2 >= 2.0.0).

`fit_curve()` and the corresponding plotting method, `plot.incidence2_fit()`
have been adapted to account for the new structure and plotting methods of
`<incidence2>` objects.

`find_peak()` is now a wrapper around the `keep_peaks()` function from
the underlying incidence2 package.

`add_rolling_average()` is no longer a generic function and will now error if
not called on an `<incidence2>` object. It now returns the original input with
additional rolling average columns but does not change the underlying objects
class. Internally the code has been refactored to make use of
`data.table::frollmean()` and function arguments updated accordingly. Due to
the changes to `add_rolling_average()`, the plot method for
`<incidence2_rolling>` objects has been removed.

`estimate_peak()` can now (optionally) return multiple peaks.

`fit_model()` is now defunct and will error if called.

Due to changes that will land in the new version of trending, I have
temporarily vendored in some of it's functionality and removed it as a
dependency. This is a temporary measure.